+++ hellolog.php
@@ -2,12 +2,12 @@
 /**
  * Plugin Name:       helloLOG
  * Plugin URI:        https://hellolog.io
- * Description:       Lightweight WordPress activity log. By hellolog.io.
+ * Description:       Lightweight WordPress activity log.
  * Version:           0.4.0
  * Requires at least: 6.4
  * Requires PHP:      8.0
- * Author:            hellolog.io
- * Author URI:        https://hellolog.io
+ * Author:            goBird
+ * Author URI:        https://gobird.io
  * License:           GPL-2.0-or-later
  * License URI:       https://www.gnu.org/licenses/gpl-2.0.html
  * Text Domain:       hellolog